requirements.txt
Django>=3.0.0<3.0.5
djangorestframework<=3.10.3<3.11.0